Fix desktop build
